:recycle: refactor (Test and Validations): Update references to model classes in helper functions
Models directly imported in helper classes and functions have been detached and dynamically referenced using Django's built-in 'get_model' method to load the model class at runtime. This was done in the context of several helper classes' functions and methods where they were being used on appointments, consent forms, cohorts, and schedules.

The reason for this change is to reduce the direct dependencies and increase the flexibility of the code. This way, if there are changes to the model itself, they will not affect the workings of these functions and methods. This is a part if broader effort to decouple model dependencies and improve code modularity and maintainability.