:recycle: refactor(CaregiverChildConsent validator): Add foreign key to CaregiverChildConsent and refactor validator
Added a new foreign key 'subject_consent' to the 'CaregiverChildConsent' model to track a subject's consent. Alongside this, the 'HIVDisclosureStatusFormValidator' logic was simplified by removing the 'child_assent' and 'maternal_delivery' dependency. Now only 'caregiver_child_consent' is used to calculate the child's age - improving efficiency and clarity of the code.